Move parsed logic to validate/parsed.js
- Validate JSON from validate.js
- Call validate/parsed from validate.js

include validate/* in lint + add parsed to index.js

return the result of validate(str)

Add test, Extend parsed from validate.parsed, add jsdoc
