+++ index.js
@@ -1,3 +1,4 @@
 exports.format = require('./lib/format');
 exports.migrate = require('./lib/migrate');
 exports.validate = require('./lib/validate');
+exports.validate.parsed = require('./validate/parsed');
