+++ package.json
@@ -8,7 +8,7 @@
     "gl-style-format": "bin/gl-style-format"
   },
   "scripts": {
-    "test": "jshint lib/*.js migrations/*.js && tape test/*.js test/migrations/*.js",
+    "test": "jshint lib/*.js migrations/*.js validate/*.js && tape test/*.js test/migrations/*.js",
     "doc": "cat lib/*.js | dox --raw --skipSingleStar | doxme > API.md",
     "cov": "istanbul cover ./node_modules/.bin/tape test/*.js test/migrations/*.js && coveralls < ./coverage/lcov.info"
   },
