+++ validate/parsed.js
@@ -1,0 +1,397 @@
+'use strict';
+
+var spec = require('mapbox-gl-style-spec');
+var parseCSSColor = require('csscolorparser').parseCSSColor;
+var format = require('util').format;
+
+/**
+ * Validate a Mapbox GL Style given a JSON object. Returns an array
+ * that can contain any number of objects representing errors. Each
+ * object has members `line` (number) if parsed was called via
+ * mapbox-gl-style-lint.validate and `message` (string).
+ *
+ * @alias parsed
+ * @param {Object} JSON a Mapbox GL Style as JSON object
+ * @returns {Array<Object>} an array of error objects
+ * @example
+ * var fs = require('fs');
+ * var validate = require('mapbox-gl-style-lint').validate;
+ * var style = JSON.parse(fs.readFileSync('./style.json', 'utf8'));
+ * var errors = validate.parsed(style);
+ */
+module.exports = function(style) {
+
+    var reference = spec['v' + style.version] || spec.latest,
+        constants = style.constants || {},
+        layers = {},
+        errors = [];
+
+    function error(key, val /*, message, ...*/) {
+        errors.push({
+            line: val.__line__,
+            message: (key ? key + ': ' : '') +
+                format.apply(format, Array.prototype.slice.call(arguments, 2))
+        });
+    }
+
+    // Main recursive validation function. Tracks:
+    //
+    // - key: string representing location of validation in style tree. Used only
+    //   for more informative error reporting.
+    // - val: current value from style being evaluated. May be anything from a
+    //   high level object that needs to be descended into deeper or a simple
+    //   scalar value.
+    // - spec: current spec being evaluated. Tracks val.
+    //
+    function validate(key, val, spec) {
+        var type = typeof_(val);
+
+        // Constants
+        if (type === 'string' && val[0] === '@') {
+            if (!(val in constants)) {
+                return error(key, val, 'constant "%s" not found', val);
+            }
+            val = constants[val];
+            type = typeof_(val);
+        }
+
+        // Functions
+        if (spec.function && type === 'object') {
+            return validate.function(key, val, spec);
+        }
+
+        if (spec.type) {
+            var validator = validate[spec.type];
+            if (validator) {
+                return validator(key, val, spec);
+            }
+            spec = reference[spec.type];
+        }
+
+        validate.object(key, val, spec);
+    }
+
+    validate.constants = function(key, val) {
+        var type = typeof_(val);
+        if (type !== 'object') {
+            return error(key, val, 'object expected, %s found', type);
+        }
+
+        for (var k in val) {
+            if (k[0] !== '@') {
+                error(key + '.' + k, val[k], 'constants must start with "@"');
+            }
+        }
+    };
+
+    validate.source = function(key, val) {
+        if (!val.type) {
+            error(key, val, '"type" is required');
+            return;
+        }
+
+        var type = unbundle(val.type);
+        switch (type) {
+            case 'vector':
+            case 'raster':
+                validate.object(key, val, reference.source_tile);
+
+                if ('url' in val) {
+                    ['tiles', 'minzoom', 'maxzoom', 'attribution'].forEach(function(prop) {
+                        if (prop in val) {
+                            error(key + '.' + prop, val[prop], 'a source with a "url" property must not include the "%s" property', prop);
+                        }
+                    });
+                }
+
+                break;
+            case 'geojson':
+                validate.object(key, val, reference.source_geojson);
+                break;
+            case 'video':
+                validate.object(key, val, reference.source_video);
+                break;
+            default:
+                validate.enum(key + '.type', val.type, {values: ['vector', 'raster', 'geojson', 'video']});
+        }
+    };
+
+    validate.layer = function(key, val) {
+        if (!val.type && !val.ref) {
+            error(key, val, 'either "type" or "ref" is required');
+        }
+
+        var type = unbundle(val.type),
+            ref = unbundle(val.ref);
+
+        if (val.id) {
+            if (layers[val.id]) {
+                error(key, val.id, 'duplicate layer id "%s", previously used at line %d', val.id, layers[val.id]);
+            } else {
+                layers[val.id] = val.id.__line__;
+            }
+        }
+
+        if ('ref' in val) {
+            ['type', 'source', 'source-layer', 'filter', 'layout'].forEach(function (p) {
+                if (p in val) {
+                    error(key, val[p], '"%s" is prohibited for ref layers', p);
+                }
+            });
+
+            var parent;
+
+            style.layers.forEach(function(layer) {
+                if (layer.id == ref) parent = layer;
+            });
+
+            if (!parent) {
+                error(key, val.ref, 'ref layer "%s" not found', ref);
+            } else if (parent.ref) {
+                error(key, val.ref, 'ref cannot reference another ref layer');
+            } else {
+                type = parent.type;
+            }
+        } else if (type !== 'background') {
+            if (!val.source) {
+                error(key, val, 'missing required property "source"');
+            } else {
+                var source = style.sources[val.source];
+                if (!source) {
+                    error(key, val.source, 'source "%s" not found', val.source);
+                } else if (source.type == 'vector' && type == 'raster') {
+                    error(key, val.source, 'layer "%s" requires a raster source', val.id);
+                } else if (source.type == 'raster' && type != 'raster') {
+                    error(key, val.source, 'layer "%s" requires a vector source', val.id);
+                }
+            }
+        }
+
+        validate.object(key, val, reference.layer, {
+            filter: validate.filter,
+            layout: function(key, val) {
+                var spec = reference['layout_' + type];
+                return type && spec && validate(key, val, spec);
+            },
+            paint: function(key, val) {
+                var spec = reference['paint_' + type];
+                return type && spec && validate(key, val, spec);
+            }
+        });
+    };
+
+    validate.object = function (key, val, spec, validators) {
+        validators = validators || {};
+
+        var type = typeof_(val);
+        if (type !== 'object') {
+            return error(key, val, 'object expected, %s found', type);
+        }
+
+        for (var k in val) {
+            var speckey = k.split('.')[0]; // treat 'paint.*' as 'paint'
+            var def = spec[speckey] || spec['*'];
+            if (def) {
+                (validators[speckey] || validate)((key ? key + '.' : key) + k, val[k], def);
+            } else {
+                if (key !== '') { // tolerate root-level extra keys
+                    error(key, val[k], 'unknown property "%s"', k);
+                }
+            }
+        }
+
+        for (var l in spec) {
+            if (spec[l].required && spec[l]['default'] === undefined && val[l] === undefined) {
+                error(key, val, 'missing required property "%s"', l);
+            }
+        }
+    };
+
+    validate.array = function (key, val, spec, validator) {
+        if (typeof_(val) !== 'array') {
+            return error(key, val, 'array expected, %s found', typeof_(val));
+        }
+
+        if (spec.length && val.length !== spec.length) {
+            return error(key, val, 'array length %d expected, length %d found', spec.length, val.length);
+        }
+
+        var value = {
+            "type": spec.value
+        };
+
+        if (style.version < 7) {
+            value.function = spec.function;
+        }
+
+        if (typeof_(spec.value) === 'object') {
+            value = spec.value;
+        }
+
+        for (var i = 0; i < val.length; i++) {
+            (validator || validate)(key + '[' + i + ']', val[i], value);
+        }
+    };
+
+    validate.filter = function(key, val) {
+        var type;
+
+        if (typeof_(val) !== 'array') {
+            return error(key, val, 'array expected, %s found', typeof_(val));
+        }
+
+        if (val.length < 1) {
+            return error(key, val, 'filter array must have at least 1 element');
+        }
+
+        validate.enum(key + '[0]', val[0], reference.filter_operator);
+
+        switch (unbundle(val[0])) {
+            case '<':
+            case '<=':
+            case '>':
+            case '>=':
+                if (val.length >= 2 && val[1] == '$type') {
+                    error(key, val, '"$type" cannot be use with operator "%s"', val[0]);
+                }
+            /* falls through */
+            case '==':
+            case '!=':
+                if (val.length != 3) {
+                    error(key, val, 'filter array for operator "%s" must have 3 elements', val[0]);
+                }
+            /* falls through */
+            case 'in':
+            case '!in':
+                if (val.length >= 2) {
+                    type = typeof_(val[1]);
+                    if (type !== 'string') {
+                        error(key + '[1]', val[1], 'string expected, %s found', type);
+                    } else if (val[1][0] === '@') {
+                        error(key + '[1]', val[1], 'filter key cannot be a constant');
+                    }
+                }
+                for (var i = 2; i < val.length; i++) {
+                    type = typeof_(val[i]);
+                    if (val[1] == '$type') {
+                        validate.enum(key + '[' + i + ']', val[i], reference.geometry_type);
+                    } else if (type === 'string' && val[i][0] === '@') {
+                        error(key + '[' + i + ']', val[i], 'filter value cannot be a constant');
+                    } else if (type !== 'string' && type !== 'number' && type !== 'boolean') {
+                        error(key + '[' + i + ']', val[i], 'string, number, or boolean expected, %s found', type);
+                    }
+                }
+                break;
+
+            case 'any':
+            case 'all':
+            case 'none':
+                for (i = 1; i < val.length; i++) {
+                    validate.filter(key + '[' + i + ']', val[i]);
+                }
+                break;
+        }
+    };
+
+    validate.function = function(key, val, spec) {
+        validate.object(key, val, reference.function, {
+            stops: function (key, val, arraySpec) {
+                var lastStop = -Infinity;
+                validate.array(key, val, arraySpec, function validateStop(key, val) {
+                    if (typeof_(val) !== 'array') {
+                        return error(key, val, 'array expected, %s found', typeof_(val));
+                    }
+
+                    if (val.length !== 2) {
+                        return error(key, val, 'array length %d expected, length %d found', 2, val.length);
+                    }
+
+                    validate(key + '[0]', val[0], {type: 'number'});
+                    validate(key + '[1]', val[1], spec);
+
+                    if (typeof_(val[0]) === 'number') {
+                        if (spec.function === 'piecewise-constant' && val[0] % 1 !== 0) {
+                            error(key + '[0]', val[0], 'zoom level for piecewise-constant functions must be an integer');
+                        }
+
+                        if (val[0] < lastStop) {
+                            error(key + '[0]', val[0], 'array stops must appear in ascending order');
+                        }
+
+                        lastStop = val[0];
+                    }
+                });
+
+                if (typeof_(val) === 'array' && val.length === 0) {
+                    error(key, val, 'array must have at least one stop');
+                }
+            }
+        });
+    };
+
+    validate.enum = function (key, val, spec) {
+        if (spec.values.indexOf(unbundle(val)) === -1) {
+            error(key, val, 'expected one of [%s], %s found', spec.values.join(', '), val);
+        }
+    };
+
+    validate.color = function(key, val) {
+        var type = typeof_(val);
+        if (type !== 'string') {
+            error(key, val, 'color expected, %s found', type);
+        } else if (parseCSSColor(val) === null) {
+            error(key, val, 'color expected, "%s" found', val);
+        }
+    };
+
+    function typeValidator(expected) {
+        return function(key, val, spec) {
+            var actual = typeof_(val);
+            if (actual !== expected) {
+                error(key, val, '%s expected, %s found', expected, actual);
+            }
+
+            if ('minimum' in spec && val < spec.minimum) {
+                error(key, val, '%s is less than the minimum value %s', val, spec.minimum);
+            }
+
+            if ('maximum' in spec && val > spec.maximum) {
+                error(key, val, '%s is greater than the maximum value %s', val, spec.maximum);
+            }
+        };
+    }
+
+    validate.number = typeValidator('number');
+    validate.string = typeValidator('string');
+    validate.boolean = typeValidator('boolean');
+
+    validate['*'] = function() {};
+
+    validate('', style, reference.$root);
+
+    return errors;
+};
+
+function typeof_(val) {
+    if (val instanceof Number)
+        return 'number';
+    if (val instanceof String)
+        return 'string';
+    if (val instanceof Boolean)
+        return 'boolean';
+    if (Array.isArray(val))
+        return 'array';
+    if (val === 'null')
+        return 'null';
+    return typeof(val);
+}
+
+function unbundle(_) {
+    if (_ instanceof Number ||
+        _ instanceof String ||
+        _ instanceof Boolean) {
+        return _.valueOf();
+    } else {
+        return _;
+    }
+}
